requirements-debian.txt
expect
screen
python3-numpy
python3-matplotlib
python3-serial
